Fix 4 more SonarQube code smells (S1185 x2, S1142, S1448)
php:S1185 x2 - ClientDwellingRepository and ProductNameTypeRepository both
had a __construct() that did nothing but call parent::__construct() with
the exact same signature — Cycle\ORM\Select\Repository's own constructor
already does this. Removed both redundant overrides; the DI factories in
config/common/di/cycle.php still work unchanged since the inherited
constructor signature is identical.

php:S1142 - CheckGatewaySecretsCommand::checkGatewaySecret() grew to 5
returns after last commit's S3776 fix (>3 allowed). Split the decode/
report branch out into decodeAndReportSecret(), leaving checkGatewaySecret()
with its original 3 (two early-exit guards + one delegated return).

php:S1448 - HomeCareSignupController grew to 21 methods (>20 allowed)
after last commit's S1142 fix added resolveDwellingAndProvisionInvoice().
Extracted the pending-signup persistence + verification-email dispatch
(savePendingSignup, processSendSignupEmail, htmlBodyWithMaskedRandomAndTimeTokenLink,
getEmailVerificationToken — 4 methods, all only reachable from signup())
into a new HomeCareSignupEmailService, following the same reasoning as
the ClientDwellingRepository/ProductNameTypeRepository splits. The
controller's MailerInterface/LoggerInterface constructor dependencies
moved with them, since nothing else in the controller used them.
Controller now at 17 methods.

Verified: php -l, full-project Psalm (no errors), full PHPUnit suite
(3896 tests, 0 failures), full Testo suite (828 tests, 0 failures).
